Automate setup so that it doesn't get done incorrectly
Also fixes install.json to at least get this loading/registering feature
Adds a placeholder for the icon because if you don't have it it explodes
