+++ README.md
@@ -2,3 +2,8 @@
 This is a cPanel plugin to flush varnish cache for user websites associated with a cPanel account.
 
 This script ensures that the cPanel plugin fetches all domains associated with the user's account and flushes the Varnish cache for each domain. Adjust the paths and commands as needed for your specific server configuration.
+
+## Installing
+`git clone` this repository, `cd` into the directory.
+Run `make`.
+Run `make uninstall` if you wish to later uninstall this.
